feat(fe): add problem detail layout
